+++ src/main/kotlin/thedarkcolour/kotlinforforge/KotlinModContainer.kt
@@ -6,11 +6,11 @@ import net.minecraftforge.eventbus.api.BusBuilder
 import net.minecraftforge.eventbus.api.Event
 import net.minecraftforge.eventbus.api.IEventBus
 import net.minecraftforge.eventbus.api.IEventListener
+import net.minecraftforge.fml.Logging
 import net.minecraftforge.fml.ModContainer
 import net.minecraftforge.fml.ModLoadingException
 import net.minecraftforge.fml.ModLoadingStage
 import net.minecraftforge.fml.event.IModBusEvent
-import net.minecraftforge.fml.loading.LogMarkers
 import net.minecraftforge.forgespi.language.IModInfo
 import net.minecraftforge.forgespi.language.ModFileScanData
 import thedarkcolour.kotlinforforge.kotlin.supply
@@ -34,7 +34,7 @@ public class KotlinModContainer(
     private val modClass: Class<*>
 
     init {
-        LOGGER.debug(LogMarkers.LOADING, "Creating KotlinModContainer instance for $className")
+        LOGGER.debug(Logging.LOADING, "Creating KotlinModContainer instance for $className")
 
         activityMap[ModLoadingStage.CONSTRUCT] = Runnable(::constructMod)
         eventBus = EventBus(BusBuilder.builder().setExceptionHandler(::onEventFailed).setTrackPhases(false).markerType(IModBusEvent::class.java))
@@ -48,9 +48,9 @@ public class KotlinModContainer(
         try {
             val layer = gameLayer.findModule(info.owningFile.moduleName()).orElseThrow()
             modClass = Class.forName(layer, className)
-            LOGGER.trace(LogMarkers.LOADING, "Loaded modclass ${modClass.name} with ${modClass.classLoader}")
+            LOGGER.trace(Logging.LOADING, "Loaded modclass ${modClass.name} with ${modClass.classLoader}")
         } catch (t: Throwable) {
-            LOGGER.error(LogMarkers.LOADING, "Failed to load class $className", t)
+            LOGGER.error(Logging.LOADING, "Failed to load class $className", t)
             throw ModLoadingException(info, ModLoadingStage.CONSTRUCT, "fml.modloading.failedtoloadmodclass", t)
         }
     }
@@ -62,21 +62,21 @@ public class KotlinModContainer(
     // Sets modInstance to a new instance of the mod class or the object instance
     private fun constructMod() {
         try {
-            LOGGER.trace(LogMarkers.LOADING, "Loading mod instance ${getModId()} of type ${modClass.name}")
+            LOGGER.trace(Logging.LOADING, "Loading mod instance ${getModId()} of type ${modClass.name}")
             modInstance = modClass.kotlin.objectInstance ?: modClass.getDeclaredConstructor().newInstance()
-            LOGGER.trace(LogMarkers.LOADING, "Loaded mod instance ${getModId()} of type ${modClass.name}")
+            LOGGER.trace(Logging.LOADING, "Loaded mod instance ${getModId()} of type ${modClass.name}")
         } catch (throwable: Throwable) {
-            LOGGER.error(LogMarkers.LOADING, "Failed to create mod instance. ModID: ${getModId()}, class ${modClass.name}", throwable)
+            LOGGER.error(Logging.LOADING, "Failed to create mod instance. ModID: ${getModId()}, class ${modClass.name}", throwable)
             throw ModLoadingException(modInfo, ModLoadingStage.CONSTRUCT, "fml.modloading.failedtoloadmod", throwable, modClass)
         }
 
         try {
-            LOGGER.trace(LogMarkers.LOADING, "Injecting Automatic Kotlin event subscribers for ${getModId()}")
+            LOGGER.trace(Logging.LOADING, "Injecting Automatic Kotlin event subscribers for ${getModId()}")
             // Inject into object EventBusSubscribers
             AutoKotlinEventBusSubscriber.inject(this, scanResults, modClass.classLoader)
-            LOGGER.trace(LogMarkers.LOADING, "Completed Automatic Kotlin event subscribers for ${getModId()}")
+            LOGGER.trace(Logging.LOADING, "Completed Automatic Kotlin event subscribers for ${getModId()}")
         } catch (throwable: Throwable) {
-            LOGGER.error(LogMarkers.LOADING, "Failed to register Automatic Kotlin subscribers. ModID: ${getModId()}, class ${modClass.name}", throwable)
+            LOGGER.error(Logging.LOADING, "Failed to register Automatic Kotlin subscribers. ModID: ${getModId()}, class ${modClass.name}", throwable)
             throw ModLoadingException(modInfo, ModLoadingStage.CONSTRUCT, "fml.modloading.failedtoloadmod", throwable, modClass)
         }
     }